classifier/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
cassandra-driver==3.29.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
DAWG2-Python==0.9.0
filelock==3.18.0
fsspec==2025.3.2
geomet==0.2.1.post1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pymorphy3==2.0.3
pymorphy3-dicts-ru==2.4.417150.4580142
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0+cpu
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.13.0
urllib3==2.3.0
